README update: clarify the site-apt preferences options
The README described a few things that were not true relating to the
apt/preferences file.

First of all it said you could ship a 'file', but preferences.pp very clearly
uses the 'content => $custom_preferences' parameter, which will not take file
sources, only templates.

Secondly, it seemed to imply that you could just drop the custom preferences
into your site-apt and it would work. But you actually need to set the
$custom_preferences to indicate the content source.

Lastly, it said that you could specify a host-specific file in the site-apt
module, but there is no facility for this (nor can you use files).

Perhaps this is where this module is going eventually, once we have a
preferences.d possibility? Until then, it makes more sense to have it reflect
the current situation.