Replaced fallback file for unattended-upgrades
with a decent one for wheezy onwards. The main syntax shouldn't
change, hopefully, from now on. It's a very basic generic one,
that uses variables for Distribution/Codename and should apply
to all coming distributions.

Removed from the default 50unattended-upgrades
----------------------------------------------

- APT::Periodic::Update-Package-Lists "1";
- APT::Periodic::Unattended-Upgrade "1";
cause this is not something the unattended upgrades class should
set by default.

- APT::UnattendedUpgrades::LogDir "/var/log/";
- APT::UnattendedUpgrades::LogFile "unattended_upgrades.log";
Because we shouldn't change the default logdest by default, which
is /var/log/unattended_upgrades/*.

- Blacklist for linux-image*, because this we shouldn't blacklist
  packages by default.